Update MergeSort.java
Added multithreading class, with implementation of Runnable, based off of the Tutorial 2 slides. It overrides run() and join() methods to include outputs every time a thread starts and finishes. As is, when the program is run, the new threads being created are displayed, so run() method seems to be working. However, I am stumped as to where to include the join() method, which has the parent threads wait for the others to finish. Playing around with it, I am only able to get the first thread to announce its completion (see comments on line 61).